Add files via upload
The CrateAway game was written for the Color Maximite 2 by vegipete. This repository is for a Sokoban level editor that I write as an adjunct to the game. This repository includes the CrateAwaySprites.png image file that was written by vegipete and used by my code, with permission by the author.